requirements.txt
streamlit
langchain
openai
streamlit-feedback
pypdf
sentence_transformers
tiktoken
python-dotenv